Tutorial on how to install Myria3D on WSL2 (#53)
* Tutorial on how to install Myria3D on WSL2

For Windows users it is quite tricky to get everything working in Myria3D. It's best to use Windows Subsystem for Linux (WSL2) to get the most out of this library if you want to work on Windows. While setting up WSL and installing Myria3D is not that difficult, if you want to utilize GPU, it becomes much more challenging. This guide tries to address the necessary steps that needs to be taken to get a properly working environment with GPU support.

* Update index.rst with new filenames

* Update and rename setup_install.md to install_on_linux.md

* Update and rename Install Myria3D on WSL2 with CUDA support.md to install_on_wsl2.md

* Update install_on_wsl2.md